+++ src/OpenConext/EngineBlockFunctionalTestingBundle/Controllers/WayfController.php
@@ -18,6 +18,8 @@
 
 namespace OpenConext\EngineBlockFunctionalTestingBundle\Controllers;
 
+use OpenConext\EngineBlock\Service\Wayf\IdpSplitter;
+use OpenConext\EngineBlockBundle\Service\WayfViewModelFactory;
 use OpenConext\EngineBlockFunctionalTestingBundle\Helper\TestEntitySeeder;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -31,11 +33,11 @@ use Twig\Environment;
  */
 class WayfController extends AbstractController
 {
-    private $twig;
-
-    public function __construct(Environment $twig)
-    {
-        $this->twig = $twig;
+    public function __construct(
+        private readonly Environment $twig,
+        private readonly IdpSplitter $idpSplitter,
+        private readonly WayfViewModelFactory $wayfViewModelFactory,
+    ) {
     }
 
     public function wayfAction(Request $request)
@@ -49,6 +51,7 @@ class WayfController extends AbstractController
         $cutoffPointForShowingUnfilteredIdps = $request->query->getInt('cutoffPointForShowingUnfilteredIdps', 100);
         $showIdPBanner = $request->query->getBoolean('showIdPBanner', true);
         $defaultIdpEntityId = $request->query->get('defaultIdpEntityId');
+        $preferredIdpEntityIds = $request->query->all('preferredIdpEntityIds');
 
         $connectedIdps = $request->query->getInt('connectedIdps', 5);
         $unconnectedIdps = $request->query->getInt('unconnectedIdps');
@@ -58,22 +61,35 @@ class WayfController extends AbstractController
             ? TestEntitySeeder::buildIdps($connectedIdps, $unconnectedIdps, $currentLocale, $defaultIdpEntityId, $addDiscoveries)
             : TestEntitySeeder::buildRandomIdps($randomIdps, $currentLocale, $defaultIdpEntityId);
 
+        $split = $this->idpSplitter->split($idpList, $preferredIdpEntityIds);
+        $preferredIdpList = $split['preferred'];
+        $regularIdpList = $split['regular'];
+        $showPreferredIdps = !empty($preferredIdpList);
+
+        $isDefaultIdpPreferred = in_array($defaultIdpEntityId, $preferredIdpEntityIds, true);
+        $showIdPBanner = $showIdPBanner && (!$showPreferredIdps || !$isDefaultIdpPreferred);
+
+        $viewModel = $this->wayfViewModelFactory->create(
+            idpList: $idpList,
+            regularIdpList: $regularIdpList,
+            preferredIdpList: $preferredIdpList,
+            showPreferredIdps: $showPreferredIdps,
+            action: $this->generateUrl('functional_testing_handle_wayf'),
+            greenHeader: $currentLocale,
+            helpLink: '/authentication/idp/help-discover?lang=' . $currentLocale,
+            backLink: $backLink,
+            cutoffPointForShowingUnfilteredIdps: $cutoffPointForShowingUnfilteredIdps,
+            showIdPBanner: $showIdPBanner,
+            rememberChoiceFeature: $rememberChoiceFeature,
+            showRequestAccess: $displayUnconnectedIdpsWayf,
+            requestId: 'bogus-request-id',
+            serviceProvider: TestEntitySeeder::buildSp(),
+            showRequestAccessContainer: true,
+        );
+
         return new Response($this->twig->render(
             '@theme/Authentication/View/Proxy/wayf.html.twig',
-            [
-                'action' => $this->generateUrl('functional_testing_handle_wayf'),
-                'greenHeader' => $currentLocale,
-                'helpLink' => '/authentication/idp/help-discover?lang='.$currentLocale,
-                'backLink' => $backLink,
-                'cutoffPointForShowingUnfilteredIdps' => $cutoffPointForShowingUnfilteredIdps,
-                'showIdPBanner' => $showIdPBanner,
-                'rememberChoiceFeature' => $rememberChoiceFeature,
-                'showRequestAccess' => $displayUnconnectedIdpsWayf,
-                'requestId' => 'bogus-request-id',
-                'serviceProvider' => TestEntitySeeder::buildSp(),
-                'idpList' => $idpList,
-                'showRequestAccessContainer' => true,
-            ]
+            $viewModel->toArray()
         ));
     }