+++ tests/library/EngineBlock/Test/Corto/Module/BindingsTest.php
@@ -20,7 +20,9 @@ use Mockery as m;
 use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
 use OpenConext\EngineBlock\Metadata\Entity\IdentityProvider;
 use OpenConext\EngineBlock\Metadata\Entity\ServiceProvider;
+use OpenConext\EngineBlock\Service\Wayf\IdpSplitter;
 use OpenConext\EngineBlockBundle\Bridge\DiContainerRuntime;
+use OpenConext\EngineBlockBundle\Service\WayfViewModelFactory;
 use PHPUnit\Framework\TestCase;
 use SAML2\Assertion;
 use SAML2\Assertion\Validation\ConstraintValidator\NotBefore;
@@ -59,7 +61,7 @@ class EngineBlock_Test_Corto_Module_BindingsTest extends TestCase
         Phake::when($this->proxyServer)->getConfig('WantsAuthnRequestsSigned')->thenReturn(false);
 
         $engineBlock = \EngineBlock_ApplicationSingleton::getInstance();
-        $engineBlock->setDiContainerRuntime(new DiContainerRuntime(Phake::mock(Twig\Environment::class)));
+        $engineBlock->setDiContainerRuntime(new DiContainerRuntime(Phake::mock(Twig\Environment::class), new IdpSplitter(), Phake::mock(WayfViewModelFactory::class)));
 
         $this->bindings = new EngineBlock_Corto_Module_Bindings($this->proxyServer);
     }